[REFACT] 빈 컴포넌트 의존성 주입방법으로 변경하여,  사용하지 않는 클래스 삭제
